Device adapters fix compiler warnings
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@17192 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/IIDC/IIDCCamera.cpp
@@ -841,7 +841,7 @@ Camera::HandleCapturedFrame(dc1394video_frame_t* frame)
 {
    PixelFormat pf = PixelFormatForLibDC1394ColorCoding(frame->color_coding);
    captureFrameCallback_(frame->image, frame->size[0], frame->size[1], pf,
-         frame->timestamp);
+         (uint32_t) frame->timestamp);
 }
